update sale
falta corregir cantidad de venta en pulg o m segun el producto y descontar los m2 ,m3, pulg, y calcular las cantidades(stock) que equivale a esos

+++ database/migrations/2019_11_04_001246_create_sold_products_table_migration.php
@@ -17,7 +17,7 @@ class CreateSoldProductsTableMigration extends Migration
             $table->bigIncrements('id');
             $table->unsignedBigInteger('sale_id');
             $table->unsignedBigInteger('product_id');
-            $table->integer('qty');
+            $table->float('qty',8,3);
             $table->decimal('price', 10, 2);
             $table->decimal('total_amount', 10, 2);
             $table->timestamps();
